+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.14.4] - 2021-10-14
+- Get Project Settings endpoint implementation
+- Implemented signature algorithm
+- Support for UI settings
+
 ## [0.14.3] - 2021-09-20
 - Fixes a bug when Client-Id/Client-Secret is not passed when using Bearer token
 - Updated domain from getkevin.eu to kevin.eu